requirements.txt
alive_progress==3.1.1
findiff==0.9.2
geopandas==0.12.2
gmsh==4.11.1
matplotlib==3.7.1
numpy==1.24.3
phidl==1.6.3
PyYAML==6.0
scipy==1.10.1
Shapely==2.0.1
tabulate==0.9.0
